backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic-settings==2.4.0
python-dotenv==1.0.1
loguru==0.7.2
sqlalchemy>=2.0.35
# psycopg2-binary==2.9.9  # PostgreSQL için - production'da gerekli
alembic==1.13.1
passlib[bcrypt]==1.7.4
bcrypt<5.0.0
python-jose[cryptography]==3.3.0
cryptography==42.0.5
requests==2.31.0
jsonpath-ng==1.6.0
weasyprint==62.3
reportlab==4.2.0
apscheduler==3.10.4
redis==5.0.1
pytest==8.2.0
pytest-asyncio==0.23.3
httpx==0.26.0
black==24.4.2
bandit==1.7.8
